INDI camera: do not assume binned frame size
